# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "GB-Locações API",
    "version": "1.0.0",
    "description": "API completa da plataforma GB-Locações para locação de equipamentos de construção civil.\n\n## Funcionalidades\n\n- 🏗️ **Equipamentos**: Catálogo completo de equipamentos para locação\n- 📋 **Categorias**: Organização por tipos de equipamento\n- 💰 **Orçamentos**: Sistema de solicitação e gestão de orçamentos\n- 🔐 **Autenticação**: Sistema de login com roles (ADMIN/CLIENT)\n- 📤 **Upload**: Sistema de upload de arquivos e imagens\n- 📧 **Contato**: Formulário de contato e comunicação\n\n## Autenticação\n\nA API utiliza JWT tokens via NextAuth.js. Para endpoints protegidos, inclua o token no header: `Authorization: Bearer <token>`\n\n### Roles Disponíveis:\n- **ADMIN**: Acesso completo a todas as funcionalidades\n- **CLIENT**: Acesso limitado a funcionalidades públicas"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    },
    {
      "url": "https://gblocacoes.vercel.app/api",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid request parameters"
                },
                "code": {
                  "type": "string",
                  "example": "400"
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "An unexpected error occurred"
                },
                "code": {
                  "type": "string",
                  "example": "500"
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/admin/quotes": {
      "get": {
        "operationId": "get-admin-quotes",
        "summary": "@openapi",
        "description": "",
        "tags": ["Admin"],
        "parameters": [],
        "responses": {
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/fix-storage": {
      "post": {
        "operationId": "post-fix-storage",
        "summary": "",
        "description": "",
        "tags": ["Fix-storage"],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        },
        "responses": {
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  }
}